+++ src/lib.rs
@@ -74,7 +74,7 @@ mod tests;
 #[cfg(feature = "serde")]
 mod serde;
 
-pub use crate::intervals::Interval;
+pub use crate::intervals::{Interval, ParseError};
 pub use crate::multi_intervals::MultiInterval;
 pub use crate::nothing_between::NothingBetween;